[DEV-1452] Refactoring

+++ src/blockchain/ain/jellyfish/services/utxo-reservation.service.ts
@@ -5,6 +5,7 @@ import { UtxoReservationRepository } from '../repositories/utxo-reservation.repo
 import { UtxoReservation } from '../domain/utxo-reservation.entity';
 import { Cron, CronExpression } from '@nestjs/schedule';
 import { Lock } from 'src/shared/lock';
+import { Config, Process } from 'src/config/config';
 
 @Injectable()
 export class UtxoReservationService implements OnModuleInit {
@@ -61,6 +62,7 @@ export class UtxoReservationService implements OnModuleInit {
   @Cron(CronExpression.EVERY_HOUR)
   @Lock()
   async cleanupOldReservations() {
+    if (Config.processDisabled(Process.CLEAN_UP_RESERVATION)) return;
     await this.repo.delete({ updated: LessThan(Util.daysBefore(7)) });
   }
 }
